Move testErrorsCaughtByApiAndRestApiWorks to its own class
This is a workaround to avoid the initialisation error described in

https://issues.apache.org/jira/browse/BROOKLYN-611
